move all wrappers to landing-page.js file
