+++ src/components/view/landing-page/faq.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  Flex,
   Box,
   Heading,
   Divider,
@@ -24,53 +23,51 @@ function Faq() {
     activeFaq === 'student' ? studentQuestions : donatorQuestions;
 
   return (
-    <Flex id="faq" bg="gray.900" py={16} px={4}>
-      <Container mt={0}>
-        <Box width="full" mb={12} textAlign="center" color="gray.100">
-          <Heading size="xl">{t('titles:FAQ')}</Heading>
-          <HStack spacing={4} my={10} justify="center">
-            {options.map(option => (
-              <Button
-                key={option}
-                onClick={() => setActiveFaq(option)}
-                isActive={activeFaq === option}
-                bg="gray.200"
-                color="gray.900"
-                _active={{ bg: 'blue.700', color: 'white' }}
-                _hover={{
-                  bg: 'gray.100',
-                  _active: { bg: 'blue.600' },
-                }}
-              >
-                {t(`I am ${option}`)}
-              </Button>
-            ))}
-          </HStack>
+    <Container mt={0}>
+      <Box width="full" mb={12} textAlign="center" color="gray.100">
+        <Heading size="xl">{t('titles:FAQ')}</Heading>
+        <HStack spacing={4} my={10} justify="center">
+          {options.map(option => (
+            <Button
+              key={option}
+              onClick={() => setActiveFaq(option)}
+              isActive={activeFaq === option}
+              bg="gray.200"
+              color="gray.900"
+              _active={{ bg: 'blue.700', color: 'white' }}
+              _hover={{
+                bg: 'gray.100',
+                _active: { bg: 'blue.600' },
+              }}
+            >
+              {t(`I am ${option}`)}
+            </Button>
+          ))}
+        </HStack>
 
-          <Divider maxW={24} borderColor="gray.100" marginX="auto" mt={8} />
-        </Box>
-        <Accordion width="full" allowMultiple>
-          <Grid
-            templateColumns={{
-              base: 'inherit',
-              md: '50% auto',
-              lg: '40% auto',
-              xl: 'repeat(2, 1fr)',
-            }}
-            width="full"
-            columnGap={8}
-            rowGap={4}
-          >
-            <QuestionList
-              questions={questionType.filter((question, i) => i % 2 === 0)}
-            />
-            <QuestionList
-              questions={questionType.filter((question, i) => i % 2 !== 0)}
-            />
-          </Grid>
-        </Accordion>
-      </Container>
-    </Flex>
+        <Divider maxW={24} borderColor="gray.100" marginX="auto" mt={8} />
+      </Box>
+      <Accordion width="full" allowMultiple>
+        <Grid
+          templateColumns={{
+            base: 'inherit',
+            md: '50% auto',
+            lg: '40% auto',
+            xl: 'repeat(2, 1fr)',
+          }}
+          width="full"
+          columnGap={8}
+          rowGap={4}
+        >
+          <QuestionList
+            questions={questionType.filter((question, i) => i % 2 === 0)}
+          />
+          <QuestionList
+            questions={questionType.filter((question, i) => i % 2 !== 0)}
+          />
+        </Grid>
+      </Accordion>
+    </Container>
   );
 }
